Informacije o korisniku i premještanje povijesti
#72
Kreiran je prikaz informacija o korisniku te je povijest korisnikovih posudba/rezervacija premještena na zaseban ekran iz razloga što se listview ne može koristiti zajedno sa scrollview-om. Razlog tome je što se listview ne može scrollat kada se nalazi unutar scrollview-a. Dodan je gumb na čiji se pritisak otvara ekran sa poviješću.
