+++ FOIKnjiznica/FOIKnjiznica/FOIKnjiznica/Classes/Clanovi.cs
@@ -10,10 +10,10 @@ namespace FOIKnjiznica.Classes
     public static class Clanovi
     {
         public static int id { get; set; } = 4;
-        public static string hrEduPersonUniqueID { get; set; }
-        public static string ime { get; set; }
-        public static string prezime { get; set; }
-        public static string mobitelID { get; set; }
+        public static string hrEduPersonUniqueID { get; set; } = "stiven@foi.hr";
+        public static string ime { get; set; } = "Stiven";
+        public static string prezime { get; set; } = "Drvoderić";
+        public static string mobitelID { get; set; } = "00012223111223";
         public static List<Publikacije> listaFavorita { get; set; }
 
         public async static void DohvatiFavorite()
