+++ FOIKnjiznica/FOIKnjiznica/FOIKnjiznica/FOIKnjiznica.csproj
@@ -55,6 +55,9 @@
     <EmbeddedResource Update="PopUpPages\SortiranjePopupPage.xaml">
       <Generator>MSBuild:UpdateDesignTimeXaml</Generator>
     </EmbeddedResource>
+    <EmbeddedResource Update="PovijestKorisnika.xaml">
+      <Generator>MSBuild:UpdateDesignTimeXaml</Generator>
+    </EmbeddedResource>
     <EmbeddedResource Update="Profil.xaml">
       <Generator>MSBuild:UpdateDesignTimeXaml</Generator>
     </EmbeddedResource>
